pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Girish Gopaul", email = "girish@gopaul.me"},
]
dependencies = [
    "fastapi>=0.111.0",
    "sqlalchemy>=2.0.31",
    "alembic>=1.13.1",
    "psycopg2>=2.9.9",
    "bcrypt>=4.1.3",
    "pyjwt>=2.8.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
